refactor: ViewController에서 View 객체 분리 및 파일 분리
