+++ BankManagerUIApp/BankManagerUIApp/View/ButtonStackView/AddClientButton.swift
@@ -1,5 +1,5 @@
 //
-//  DoingTaskLabel.swift
+//  AddClientButton.swift
 //  BankManagerUIApp
 //
 //  Created by 리지, 무리 on 2023/03/14.
@@ -7,7 +7,7 @@
 
 import UIKit
 
-final class DoingTaskLabel: UILabel {
+final class AddClientButton: UIButton {
     override init(frame: CGRect) {
         super.init(frame: frame)
         configure()
@@ -17,12 +17,10 @@ final class DoingTaskLabel: UILabel {
         fatalError("init(coder:) has not been implemented")
     }
     
-    func configure() {
+    private func configure() {
         self.translatesAutoresizingMaskIntoConstraints = false
-        self.textAlignment = .center
-        self.font = UIFont.preferredFont(forTextStyle: .title1)
-        self.text = "업무중"
-        self.textColor = .white
-        self.backgroundColor = .systemIndigo
+        self.setTitle("고객 10명 추가", for: .normal)
+        self.setTitleColor(.systemBlue, for: .normal)
+        self.titleLabel?.font = UIFont.preferredFont(forTextStyle: .callout)
     }
 }
