+++ BankManagerUIApp/BankManagerUIApp/View/QueueStackView/DoingTaskLabel.swift
@@ -1,5 +1,5 @@
 //
-//  ResetButton.swift
+//  DoingTaskLabel.swift
 //  BankManagerUIApp
 //
 //  Created by 리지, 무리 on 2023/03/14.
@@ -7,7 +7,7 @@
 
 import UIKit
 
-final class ResetButton: UIButton {
+final class DoingTaskLabel: UILabel {
     override init(frame: CGRect) {
         super.init(frame: frame)
         configure()
@@ -17,10 +17,12 @@ final class ResetButton: UIButton {
         fatalError("init(coder:) has not been implemented")
     }
     
-    func configure() {
+    private func configure() {
         self.translatesAutoresizingMaskIntoConstraints = false
-        self.setTitle("초기화", for: .normal)
-        self.setTitleColor(.systemRed, for: .normal)
-        self.titleLabel?.font = UIFont.preferredFont(forTextStyle: .body)
+        self.textAlignment = .center
+        self.font = UIFont.preferredFont(forTextStyle: .largeTitle)
+        self.text = "업무중"
+        self.textColor = .white
+        self.backgroundColor = .systemIndigo
     }
 }
