+++ BankManagerUIApp/BankManagerUIApp/View/TaskTimerLabel.swift
@@ -1,5 +1,5 @@
 //
-//  WaitingQueueLabel.swift
+//  TaskTimerLabel.swift
 //  BankManagerUIApp
 //
 //  Created by 리지, 무리 on 2023/03/14.
@@ -7,7 +7,7 @@
 
 import UIKit
 
-final class WaitingQueueLabel: UILabel {
+final class TaskTimerLabel: UILabel {
     override init(frame: CGRect) {
         super.init(frame: frame)
         configure()
@@ -16,13 +16,11 @@ final class WaitingQueueLabel: UILabel {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
-    func configure() {
+   
+    private func configure() {
         self.translatesAutoresizingMaskIntoConstraints = false
         self.textAlignment = .center
-        self.font = UIFont.preferredFont(forTextStyle: .title1)
-        self.text = "대기중"
-        self.textColor = .white
-        self.backgroundColor = .systemGreen
+        self.font = UIFont.preferredFont(forTextStyle: .title2)
+        self.text = "업무시간 - "
     }
 }
